backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
alembic
redis
asyncpg
openai>=1.40.0
yfinance
tavily-python

